[refactor] factors drone view out
